+++ cmd/neofs-node/config.go
@@ -313,8 +313,6 @@ type cfgLocalStorage struct {
 }
 
 type cfgObjectRoutines struct {
-	putRemote *ants.Pool
-
 	search *ants.Pool
 }
 
@@ -591,9 +589,6 @@ func initObjectPool(cfg *config.Config) (pool cfgObjectRoutines) {
 
 	optNonBlocking := ants.WithNonblocking(true)
 
-	pool.putRemote, err = ants.NewPool(cfg.Object.Put.PoolSizeRemote, optNonBlocking)
-	fatalOnErr(err)
-
 	pool.search, err = ants.NewPool(cfg.Object.Search.PoolSize, optNonBlocking)
 	fatalOnErr(err)
 
@@ -603,8 +598,6 @@ func initObjectPool(cfg *config.Config) (pool cfgObjectRoutines) {
 func (c *cfg) reloadObjectPoolSizes() {
 	c.cfgObject.poolLock.Lock()
 	defer c.cfgObject.poolLock.Unlock()
-
-	c.cfgObject.pool.putRemote.Tune(c.appCfg.Object.Put.PoolSizeRemote)
 
 	c.cfgObject.pool.search.Tune(c.appCfg.Object.Search.PoolSize)
 }