+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Changelog for NeoFS Node
 - Proxied GET data is not checked against SHA256 hash from header (#4081)
 - SN requests EC part ranges using GET API instead of deprecated RANGE one (#4067)
 - In-container SN no longer attaches session/bearer tokens to spawned GET/HEAD requests (#4067)
+- SN no longer limits number of concurrent outgoing requests in PUT handler (#4092)
 
 ### Removed
 - Compression support from FSTree (#4054)
@@ -38,6 +39,8 @@ IR container structurization migration introduced in v0.51.0 has been removed.
 Inner Ring nodes updating from versions earlier than v0.51.0 must first update
 to a v0.51.0-v0.54.0 release and let the migration complete before updating to
 v0.55.0.
+
+Drop no longer supported `object.put.pool_size_remote` option from all SN configurations.
 
 ## [0.54.0] - 2026-06-29 - Ando